clean & adding uart dma

+++ src/grbl/grbl_src/serial.h
@@ -54,11 +54,11 @@ uint8_t serial_get_rx_buffer_available();
 
 // Returns the number of bytes used in the RX serial buffer.
 // NOTE: Deprecated. Not used unless classic status reports are enabled in config.h.
-uint8_t serial_get_rx_buffer_count();
+uint16_t serial_get_rx_buffer_count();
 
 // Returns the number of bytes used in the TX serial buffer.
 // NOTE: Not used except for debugging and ensuring no TX bottlenecks.
-uint8_t serial_get_tx_buffer_count();
+uint16_t serial_get_tx_buffer_count();
 
 // 串口中断执行的函数
 void laser_uart_rx_handler(__IO uint8_t data);
